backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
anthropic==0.39.0
pytest==8.3.3
httpx==0.27.2
